refactored examples with new folder structure
