--- mat/examples/ble/simple/_macs.py
@@ -1,7 +1,0 @@
-my_ = '04:ee:03:6c:ef:49'   # cc26x2, mine, red wire
-ad7 = '04:ee:03:6c:ef:30'   # cc26x2, ad7788
-puz = '04:ee:03:6c:ef:e0'   # cc26x2, purple zip-tie
-nop = 'ff:ff:ff:ff:ff:ff'   # fake one
-sxt037 = '04:ee:03:73:87:43'
-sxt050 = '04:ee:03:73:87:45'
-lp2 = '80:6f:b0:1e:3d:18'
