+++ mat/examples_ble/_sn.py
@@ -1,0 +1,17 @@
+# template
+# _pre = 'YYYY'
+# _sn = 'XXX'
+# _sn_full = '{}{}'.format(_pre, _sn)
+# _sn_to_mac = {
+#     'AAA': 'FF:FF:FF:FF:FF:FF'
+# }
+
+
+# real, grab client file from utils_lowell
+_pre = 'YYYY'
+_sn = 'XXX'
+_sn_full = '{}{}'.format(_pre, _sn)
+_sn_to_mac = {
+    'AAA': 'FF:FF:FF:FF:FF:FF'
+}
+mac = _sn_to_mac[_sn]
