+++ mat/examples_ble/recipes/exe_fr10s.sh
