+++ mat/examples_ble/simple/_bluepy.py
@@ -1,6 +1,9 @@
 import bluepy.btle as ble
 from mat.logger_controller_ble import LoggerControllerBLE
-from mat.examples.ble.simple._macs import mac
+from mat.examples_ble._macs import _macs
+
+# used in these examples
+mac = _macs['lp2']
 
 
 def signal_strength(a=mac):
