+++ mat/examples_ble/simple/del.py
@@ -2,7 +2,10 @@ import bluepy.btle as ble
 from mat.logger_controller import (
     DEL_FILE_CMD)
 from mat.logger_controller_ble import LoggerControllerBLE
-from mat.examples.ble.simple._macs import mac
+from mat.examples_ble._macs import _macs
+
+# used in these examples
+mac = _macs['lp2']
 
 
 def rm(f):
