+++ mat/examples_ble/simple/gtm_stm.py
@@ -1,11 +1,10 @@
 import time
-
 import bluepy.btle as ble
 from mat.logger_controller_ble import LoggerControllerBLE
-from mat.examples.ble.simple import _macs
+from mat.examples_ble._macs import _macs
 
-# mac = _macs.puz
-mac = _macs.sxt050
+# used in these examples
+mac = _macs['lp2']
 
 
 def gtm_stm():
