+++ mat/examples_ble/simple/log.py
@@ -1,9 +1,9 @@
 import bluepy.btle as ble
 from mat.logger_controller_ble import LoggerControllerBLE, LOG_EN_CMD
-from mat.examples.ble.simple import _macs
+from mat.examples_ble._macs import _macs
 
-# mac = _macs.puz
-mac = _macs.sxt050
+# used in these examples
+mac = _macs['lp2']
 
 
 def log_en():
