+++ mat/examples_ble/simple/ls_dwl.py
@@ -1,8 +1,11 @@
 import time
 import bluepy.btle as ble
 from mat.logger_controller_ble import LoggerControllerBLE
-from mat.examples.ble.simple._macs import mac
 from pprint import pprint
+from mat.examples_ble._macs import _macs
+
+# used in these examples
+mac = _macs['lp2']
 
 
 # usually, this creates a big file for testing
