+++ mat/examples_ble/simple/mts.py
@@ -1,24 +1,21 @@
 import bluepy.btle as ble
-from mat.logger_controller import REQ_FILE_NAME_CMD
 from mat.logger_controller_ble import LoggerControllerBLE
-from mat.examples.ble.simple import _macs
+from mat.examples_ble._macs import _macs
 
-# mac = _macs.lp2
-mac = _macs.sxt050
+# used in these examples
+mac = _macs['lp2']
 
 
-def rfn():
+def mts():
     try:
         with LoggerControllerBLE(mac) as lc_ble:
-            result = lc_ble.command(REQ_FILE_NAME_CMD)
-            print('\t\tRFN --> {}'.format(result))
+            result = lc_ble.command('MTS')
+            print('\t\tMTS --> {}'.format(result))
     except ble.BTLEException as ex:
         print('BLE: connect exception --> {}'.format(ex))
 
 
 if __name__ == '__main__':
     print('APP: start')
-
-    rfn()
-
+    mts()
     print('APP: done')
