+++ mat/examples_ble/simple/rfn.py
@@ -1,21 +1,24 @@
 import bluepy.btle as ble
+from mat.logger_controller import REQ_FILE_NAME_CMD
 from mat.logger_controller_ble import LoggerControllerBLE
-from mat.examples.ble.simple import _macs
+from mat.examples_ble._macs import _macs
 
-# mac = _macs.puz
-mac = _macs.sxt050
+# used in these examples
+mac = _macs['lp2']
 
 
-def mts():
+def rfn():
     try:
         with LoggerControllerBLE(mac) as lc_ble:
-            result = lc_ble.command('MTS')
-            print('\t\tMTS --> {}'.format(result))
+            result = lc_ble.command(REQ_FILE_NAME_CMD)
+            print('\t\tRFN --> {}'.format(result))
     except ble.BTLEException as ex:
         print('BLE: connect exception --> {}'.format(ex))
 
 
 if __name__ == '__main__':
     print('APP: start')
-    mts()
+
+    rfn()
+
     print('APP: done')
